feat(backend): remove config from graphql state
